revert: inline test config back into test files, remove session_test_config.py
